requirements.txt
apify-client>=1.6.4
pandas>=2.1.2
streamlit>=1.28.1
plotly>=5.17.0
gspread>=5.12.0
google-auth>=2.23.3
python-dotenv>=1.0.0
pytz>=2023.3
openpyxl>=3.1.2
requests>=2.31.0
